Update alltypes.h to match reference-sysroot types too.
This includes off_t, ino_t, and several other POSIX types.

+++ system/lib/libc/musl/arch/emscripten/bits/alltypes.h
@@ -35,14 +35,14 @@ typedef __WCHAR_TYPE__ wchar_t;
 
 #else
 #if defined(__NEED_wchar_t) && !defined(__DEFINED_wchar_t)
-typedef long wchar_t;
+typedef int wchar_t;
 #define __DEFINED_wchar_t
 #endif
 
 #endif
 #endif
 #if defined(__NEED_wint_t) && !defined(__DEFINED_wint_t)
-typedef unsigned wint_t;
+typedef int wint_t;
 #define __DEFINED_wint_t
 #endif
 
@@ -219,17 +219,17 @@ typedef unsigned _Reg nlink_t;
 #endif
 
 #if defined(__NEED_off_t) && !defined(__DEFINED_off_t)
-typedef int off_t;
+typedef long long off_t;
 #define __DEFINED_off_t
 #endif
 
 #if defined(__NEED_ino_t) && !defined(__DEFINED_ino_t)
-typedef unsigned int ino_t;
+typedef unsigned long long ino_t;
 #define __DEFINED_ino_t
 #endif
 
 #if defined(__NEED_dev_t) && !defined(__DEFINED_dev_t)
-typedef unsigned int dev_t;
+typedef unsigned long long dev_t;
 #define __DEFINED_dev_t
 #endif
 
@@ -239,22 +239,22 @@ typedef long blksize_t;
 #endif
 
 #if defined(__NEED_blkcnt_t) && !defined(__DEFINED_blkcnt_t)
-typedef int blkcnt_t;
+typedef long long blkcnt_t;
 #define __DEFINED_blkcnt_t
 #endif
 
 #if defined(__NEED_fsblkcnt_t) && !defined(__DEFINED_fsblkcnt_t)
-typedef unsigned int fsblkcnt_t;
+typedef unsigned long long fsblkcnt_t;
 #define __DEFINED_fsblkcnt_t
 #endif
 
 #if defined(__NEED_fsfilcnt_t) && !defined(__DEFINED_fsfilcnt_t)
-typedef unsigned int fsfilcnt_t;
+typedef unsigned long long fsfilcnt_t;
 #define __DEFINED_fsfilcnt_t
 #endif
 
 #if defined(__NEED_wint_t) && !defined(__DEFINED_wint_t)
-typedef unsigned wint_t;
+typedef int wint_t;
 #define __DEFINED_wint_t
 #endif
